ai-service/requirements.txt
Flask
flask-cors
scikit-learn
pandas
joblib
python-dotenv